Added FPGA interface and debugged Phase 1
The FPGA interface, which consists of the Advanced eXtensible Interface (AXI) protocol defined by ARM, has been implemented. This is essential to porting our design onto the OCT FPGA.

A bug has also been resolved in the Position Ring Node for Phase 1, as the addresses of particles passed on throughout the pipeline doesn't include the double buffering extension.